refactor: rename the folder containing the media players (use plural form)
